requirements.txt
pandas
numpy
scikit-learn
joblib
requests
streamlit
hopsworks
python-dotenv
matplotlib
shap
lime
joblib
